requirements.txt
discord.py
python-dotenv
numpy
joblib
sentence-transformers
requests
transformers